move some mobile base utils as a seperate module
